+++ src/backend/marsha/core/tests/utils/medialive_utils/test_medialive_delete_utils.py
@@ -1,6 +1,9 @@
 """Test medialive utils functions."""
+from datetime import datetime, timezone
+from io import StringIO
 from unittest import mock
 
+from django.core.management.base import OutputWrapper
 from django.test import TestCase, override_settings
 
 from botocore.stub import Stubber
@@ -125,3 +128,129 @@ class MediaLiveUtilsTestCase(TestCase):
             deleted_endpoints = medialive_utils.delete_mediapackage_channel("1")
             mediapackage_client_stubber.assert_no_pending_responses()
         self.assertEqual(deleted_endpoints, ["1", "2"])
+
+    def test_delete_medialive_stack_on_running_channel(self):
+        """Should stop the medialive channel then delete the stack."""
+        out = StringIO()
+        live_to_delete = {
+            "Name": "dev-bar_99d60314-20f2-4847-84a4-d2f47bf7fe38_1629982356",
+            "Tags": {"environment": "test"},
+            "Id": "562345",
+            "State": "Running",
+            "InputAttachments": [
+                {
+                    "InputId": "876294",
+                }
+            ],
+        }
+
+        with mock.patch(
+            "django.utils.timezone.now",
+            return_value=datetime(2021, 8, 26, 13, 25, tzinfo=timezone.utc),
+        ), Stubber(
+            medialive_utils.medialive_client
+        ) as mediapackage_client_stubber, mock.patch.object(
+            medialive_utils.medialive_delete_utils, "delete_mediapackage_channel"
+        ) as delete_mediapackage_channel_mock:
+            mediapackage_client_stubber.add_response(
+                "stop_channel",
+                expected_params={"ChannelId": "562345"},
+                service_response={},
+            )
+            mediapackage_client_stubber.add_response(
+                "describe_channel",
+                expected_params={"ChannelId": "562345"},
+                service_response={"State": "IDLE"},
+            )
+            mediapackage_client_stubber.add_response(
+                "delete_channel",
+                expected_params={"ChannelId": "562345"},
+                service_response={},
+            )
+
+            mediapackage_client_stubber.add_response(
+                "describe_input",
+                expected_params={"InputId": "876294"},
+                service_response={"State": "DETACHED"},
+            )
+            mediapackage_client_stubber.add_response(
+                "delete_input",
+                expected_params={"InputId": "876294"},
+                service_response={},
+            )
+
+            medialive_utils.delete_medialive_stack(live_to_delete, OutputWrapper(out))
+            delete_mediapackage_channel_mock.assert_called_once_with(
+                "dev-bar_99d60314-20f2-4847-84a4-d2f47bf7fe38_1629982356"
+            )
+        self.assertEqual(
+            (
+                "Cleaning stack with name "
+                "dev-bar_99d60314-20f2-4847-84a4-d2f47bf7fe38_1629982356\n"
+                "Medialive channel is running, we must stop it first.\n"
+                "Stack with name "
+                "dev-bar_99d60314-20f2-4847-84a4-d2f47bf7fe38_1629982356 deleted\n"
+            ),
+            out.getvalue(),
+        )
+        out.close()
+
+    def test_delete_medialive_stack_on_stopped_channel(self):
+        """Should delete a medialive stack directly."""
+        out = StringIO()
+        live_to_delete = {
+            "Name": "dev-bar_99d60314-20f2-4847-84a4-d2f47bf7fe38_1629982356",
+            "Tags": {"environment": "test"},
+            "Id": "562345",
+            "State": "STOPPED",
+            "InputAttachments": [
+                {
+                    "InputId": "876294",
+                }
+            ],
+        }
+
+        with mock.patch(
+            "django.utils.timezone.now",
+            return_value=datetime(2021, 8, 26, 13, 25, tzinfo=timezone.utc),
+        ), Stubber(
+            medialive_utils.medialive_client
+        ) as mediapackage_client_stubber, mock.patch.object(
+            medialive_utils.medialive_delete_utils, "delete_mediapackage_channel"
+        ) as delete_mediapackage_channel_mock:
+            mediapackage_client_stubber.add_response(
+                "delete_channel",
+                expected_params={"ChannelId": "562345"},
+                service_response={},
+            )
+            mediapackage_client_stubber.add_response(
+                "describe_input",
+                expected_params={"InputId": "876294"},
+                service_response={"State": "DETACHED"},
+            )
+            mediapackage_client_stubber.add_response(
+                "delete_input",
+                expected_params={"InputId": "876294"},
+                service_response={},
+            )
+
+            mediapackage_client_stubber.add_response(
+                "delete_input",
+                expected_params={"InputId": "876294"},
+                service_response={},
+            )
+
+            medialive_utils.delete_medialive_stack(live_to_delete, OutputWrapper(out))
+            delete_mediapackage_channel_mock.assert_called_once_with(
+                "dev-bar_99d60314-20f2-4847-84a4-d2f47bf7fe38_1629982356"
+            )
+        self.assertEqual(
+            (
+                "Cleaning stack with name "
+                "dev-bar_99d60314-20f2-4847-84a4-d2f47bf7fe38_1629982356\n"
+                "Stack with name "
+                "dev-bar_99d60314-20f2-4847-84a4-d2f47bf7fe38_1629982356 deleted\n"
+            ),
+            out.getvalue(),
+        )
+        out.close()